Migrate errorLogService to TypeScript
